API-299: fix tests, move files, linter fixes
