+++ apps/api/src/endpoints/interaction/interaction.controller.ts
@@ -1,13 +1,13 @@
 import { Controller, Post, Body } from '@nestjs/common';
 import { ApiTags, ApiResponse, ApiOkResponse } from '@nestjs/swagger';
-import { GovernanceCreateProposalRequest } from '@libs/entities/entities/governance.create.proposal.request';
-import { TransactionDetails } from '@libs/entities/entities/transaction.details';
+import { GovernanceCreateProposalRequest } from '@libs/entities/governance.create.proposal.request';
+import { TransactionDetails } from '@libs/entities/transaction.details';
 import { InteractionService } from '@libs/services/interactions';
-import { GovernanceCloseProposalRequest } from '@libs/entities/entities/governance.close.proposal.request';
-import { GovernanceClearProposalsRequest } from '@libs/entities/entities/governance.clear.proposals.request';
-import { GovernanceChangeConfigRequest } from '@libs/entities/entities/governance.change.config.request';
-import { GovernanceVoteRequest } from '@libs/entities/entities/governance.vote.request';
-import { GovernanceDelegateVoteRequest } from '@libs/entities/entities/governance.delegate.vote.request';
+import { GovernanceCloseProposalRequest } from '@libs/entities/governance.close.proposal.request';
+import { GovernanceClearProposalsRequest } from '@libs/entities/governance.clear.proposals.request';
+import { GovernanceChangeConfigRequest } from '@libs/entities/governance.change.config.request';
+import { GovernanceVoteRequest } from '@libs/entities/governance.vote.request';
+import { GovernanceDelegateVoteRequest } from '@libs/entities/governance.delegate.vote.request';
 import { CreateProposalValidationPipe } from '@libs/pipes';
 import { VoteValidationPipe } from '@libs/pipes';
 import { DelegateVoteValidationPipe } from '@libs/pipes/delegate.vote.validation.pipe';
