--- libs/entities/src/entities/index.ts
@@ -1,6 +1,0 @@
-export * from './governance.config';
-export * from './governance.create.proposal.request';
-export * from './transaction.details';
-export * from './governance.delegated.vote.info';
-export * from './governance.proposal';
-export * from './governance.voting.power';
