+++ libs/entities/src/governance.clear.proposals.request.ts
