+++ libs/entities/src/governance.close.proposal.request.ts
