+++ libs/entities/src/governance.config.ts
