+++ libs/entities/src/governance.delegated.vote.info.ts
