+++ libs/entities/src/governance.proposal.ts
