+++ libs/entities/src/governance.voting.power.ts
