+++ libs/entities/src/index.ts
@@ -1,1 +1,11 @@
-export * from './entities';
+export * from './governance.config';
+export * from './governance.create.proposal.request';
+export * from './transaction.details';
+export * from './governance.delegated.vote.info';
+export * from './governance.proposal';
+export * from './governance.voting.power';
+export * from './governance.change.config.request';
+export * from './governance.close.proposal.request';
+export * from './governance.clear.proposals.request';
+export * from './governance.vote.request';
+export * from './vote.option';
