+++ libs/entities/src/vote.option.ts
