+++ libs/services/src/index.ts
@@ -1,4 +1,1 @@
-export * from './example';
 export * from './services.module';
-export * from './token';
-export * from './user';
