--- libs/services/src/test/jest.common.config.json
@@ -1,18 +1,0 @@
-{
-  "moduleFileExtensions": [
-    "js",
-    "json",
-    "ts"
-  ],
-  "rootDir": ".",
-  "testEnvironment": "node",
-  "testRegex": ".spec.ts$",
-  "transform": {
-    "^.+\\.(t|j)s$": "ts-jest"
-  },
-  "moduleNameMapper": {
-    "^@libs/common(|/.*)$": "<rootDir>/../../../../libs/common/src/$1",
-    "^@libs/common": "<rootDir>/../../../../libs/common",
-    "^@libs/entities": "<rootDir>/../../../../libs/entities/src"
-  }
-}
